Rename /viewmodel -> /viewmodels
