+++ app/src/main/java/com/ysshin/fine_dust_app/viewmodels/LoginViewModel.kt
@@ -1,4 +1,4 @@
-package com.ysshin.fine_dust_app.viewmodel
+package com.ysshin.fine_dust_app.viewmodels
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
